fix(display): widen map key to uint64, add artifact appearance path to .disp
The Skull of the Man'ari (item 128943) could not have its appearance
changed via .disp because the custom display handler used a hardcoded
76-entry lookup table limited to modifier IDs 0-32, and passed the
bonus parameter as uint8.

Changes:
- DB2Stores.cpp: widen _itemModifiedAppearancesByItem key from uint32
  (24-bit item ID + 8-bit modifier) to uint64 (32-bit each), preventing
  hash collisions for high modifier IDs
- custom_display_handler: uint32 bonus param, remove hardcoded mod list,
  add artifact-aware path that routes through ArtifactAppearance system
  (SetAppearanceModId + ITEM_MODIFIER_ARTIFACT_APPEARANCE_ID) before
  falling back to standard ItemModifiedAppearance lookup
- Handle paired artifact child items (offhand follows mainhand appearance)

+++ src/server/game/DataStores/DB2Structure.h
@@ -221,7 +221,7 @@ struct ArtifactAppearanceEntry
     uint16 ArtifactAppearanceSetID;
     uint8 DisplayIndex;
     uint32 UnlockPlayerConditionID;
-    uint32 ItemAppearanceModifierID;
+    uint8 ItemAppearanceModifierID;
     int32 UiSwatchColor;
     float UiModelSaturation;
     float UiModelOpacity;
